+++ sw/libs/util/inc/wav.h
@@ -33,12 +33,12 @@ typedef struct {
   WAVFileMode mode;
 
   bool eof;
-	int position;
+	uint16_t position;
 } WAVFile;
 
 void WAV_Open(WAVFile *file, char *filename);
-void WAV_Read(WAVFile *file, void *buffer, uint32_t bytesToRead);
-void WAV_Write(WAVFile *file, void *buffer, uint32_t bytesToWrite);
+void WAV_Read(WAVFile *file, void *buffer, uint16_t bytesToRead);
+void WAV_Write(WAVFile *file, void *buffer, uint16_t bytesToWrite);
 void WAV_Close(WAVFile *file);
 
 bool WAV_EOF(WAVFile *file);
